fix: cp doc_release.sh to a temp loc before swich branch

+++ .github/scripts/doc_release.sh
@@ -16,6 +16,29 @@
 # limitations under the License.
 #===============================================================================
 
+# Set git auth for push changes 
+git config --global user.name "GitHub Actions"
+git config --global user.email "41898282+github-actions[bot]@users.noreply.github.com"
+
+# Ensure the build directory exists
+BUILD_DIR="doc/_build/scikit-learn-intelex"
+if [ ! -d "$BUILD_DIR" ]; then
+    echo "::error: Documentation build directory not found!"
+    exit 1
+fi
+
+# Copy built documentation to a temp location
+DEPLOY_DIR="/tmp/gh-pages-deploy"
+mkdir -p "$DEPLOY_DIR"
+cp -R "$BUILD_DIR"/* "$DEPLOY_DIR"
+ls $DEPLOY_DIR
+
+# Checkout gh-pages branch
+if ! git checkout gh-pages; then
+    echo "::error:: Could not checkout gh-pages branch!"
+    exit 1
+fi
+
 # Move the new versioned folder to the correct location
 rm -Rf latest
 cp -R "$DEPLOY_DIR/$SHORT_DOC_VERSION" "$SHORT_DOC_VERSION"
